Add new scripts for patient data processing and model evaluation
- Introduced `per_patient_table.py` to merge model predictions, FEV1/FVC measurements, and quantitative CT indices into a comprehensive patient comparison table.
- Created `pft_new_readings.py` to extract FEV1/FVC and demographic data from new PFT report pages.
- Developed `quantitative_ct_features.py` to compute classical quantitative CT emphysema indices for each patient, including %LAA-950 and PD15.
- Implemented `repeated_nested_cv_probe.py` for repeated nested cross-validation of a frozen-embedding linear probe, enhancing model evaluation robustness.

+++ regression/core/evaluator.py
@@ -249,9 +249,12 @@ def get_regression_predictions(
             x = _extract_model_input(batch, device)
             target = _extract_regression_target(batch).to(device, non_blocking=True)
 
+            # bf16 for the same reason as in the trainer — and this path has
+            # no full-precision retry, so an fp16 NaN here silently becomes an
+            # invalid prediction rather than a caught error.
             with torch.autocast(
                 device_type=device.type,
-                dtype=torch.float16,
+                dtype=torch.bfloat16,
                 enabled=bool(use_amp and device.type == "cuda"),
             ):
                 preds = _extract_regression_predictions(model(x))
@@ -296,9 +299,12 @@ def get_classification_predictions(
             x = _extract_model_input(batch, device)
             target = _extract_classification_target(batch).to(device, non_blocking=True)
 
+            # bf16 for the same reason as in the trainer — and this path has
+            # no full-precision retry, so an fp16 NaN here silently becomes an
+            # invalid prediction rather than a caught error.
             with torch.autocast(
                 device_type=device.type,
-                dtype=torch.float16,
+                dtype=torch.bfloat16,
                 enabled=bool(use_amp and device.type == "cuda"),
             ):
                 logits = _extract_classification_logits(model(x))
